build(semantic-ui): Use semantic-ui v2.2.6
When running 'npm install', semantic.json specifies new version (2.2.6).
This happens even when running 'npm install --save-exact', when my
package.json specified the previous version (2.2.4).

The release notes on this package do not concern any parts of
semantic-ui that I am using at the moment. Nonetheless this upgrade is
an annoyance, so just go with it.