test(utils): Improve tests for areArraysEqual
Add a helper method to make for easier testing. It's important that
this method makes a clone of the incoming data. But there was no easy
way to test this, since the function returns a boolean. The helper
method takes care of this important side effect, now I can just test
that the method was called.